+++ tests/test_profile.py
@@ -63,10 +63,9 @@ def test_profile_if_there_is_no_information_on_collected_tasks(tmp_path, runner)
 def test_profile_if_there_is_information_on_collected_tasks(tmp_path, runner):
     source = """
     import time
-    import pytask
+    from pathlib import Path
 
-    @pytask.mark.produces("out.txt")
-    def task_example(produces):
+    def task_example(produces=Path("out.txt")):
         time.sleep(2)
         produces.write_text("There are nine billion bicycles in Beijing.")
     """